requirements.txt
mongoengine==0.17.0
aiogram==2.3
aiohttp==3.6.1
aiohttp-socks==0.2.2
aiohttp-swagger==1.0.9
pyyaml==5.1
sentry-sdk==0.12.3
dnspython
pandas==0.25.1
openpyxl==3.0.0
xlrd==1.2.0
py-rouge==1.1
nltk==3.2.5
aio-pika==5.6.0
